+++ OurUmbraco.Site/OurUmbraco.Site.csproj
@@ -4355,6 +4355,7 @@
     <Content Include="Views\CommunityStatistics.cshtml" />
     <Content Include="Views\CommunityHubPage.cshtml" />
     <Content Include="Views\CommunityHubPageLayout.cshtml" />
+    <Content Include="Views\CommunityKarma.cshtml" />
     <None Include="web.Debug.config">
       <DependentUpon>web.config</DependentUpon>
     </None>
