+++ OurUmbraco.Site/Views/CommunityKarma.cshtml
@@ -1,0 +1,6 @@
+﻿@inherits OurUmbraco.Our.Models.OurUmbracoTemplatePage
+@{
+    Layout = "~/Views/CommunityHubPageLayout.cshtml";
+}
+
+@Html.Partial("~/Views/Partials/Community/KarmaStatistics.cshtml")
