+++ OurUmbraco/OurUmbraco.csproj
@@ -396,6 +396,7 @@
     <Compile Include="Community\GitHub\Models\GithubPullRequestModel.cs" />
     <Compile Include="Community\GitHub\Models\IGitHubContributorModel.cs" />
     <Compile Include="Community\GitHub\Models\GitHubContributorsResult.cs" />
+    <Compile Include="Community\Karma\KarmaService.cs" />
     <Compile Include="Community\Meetup\MeetupService.cs" />
     <Compile Include="Community\Meetup\Models\MeetupCacheItem.cs" />
     <Compile Include="Community\Models\MeetupEventsModel.cs" />
